rename KeyedPartition to StaticPartitionMap add docs for StaticPartitionMap fix code broken by suggestion
